Remove all the things! This is a rewrite
Joel says this is the absolute worst thing you can do, because all of
the cruft of your old codebase is actually working code.

If this was true for the prior helix pi codebase maybe I would be more
liberal. This was my first truely major JavaScript project. It barely
worked some of the time, suffered fundamental shared mutable reference
bugs, and was generally lacking in quality design.

So away it goes.